requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
passlib==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
